:art: Move CoreOptionsExtension and QueryTrackingBehavior
