+++ src/EntityFramework.Core/EntityFramework.Core.csproj
@@ -67,7 +67,6 @@
     <Compile Include="ChangeTracking\Internal\SimpleKeyValueFactory.cs" />
     <Compile Include="ChangeTracking\Internal\ValueBufferOriginalValues.cs" />
     <Compile Include="ChangeTracking\Internal\ValueBufferSidecar.cs" />
-    <Compile Include="ChangeTracking\QueryTrackingBehavior.cs" />
     <Compile Include="DbUpdateConcurrencyException.cs" />
     <Compile Include="DbUpdateException.cs" />
     <Compile Include="Extensions\Internal\CoreLoggerExtensions.cs" />
@@ -81,6 +80,7 @@
     <Compile Include="Infrastructure\IKeyValue.cs" />
     <Compile Include="Infrastructure\ModelSource.cs" />
     <Compile Include="Infrastructure\CoreLoggingEventId.cs" />
+    <Compile Include="Internal\CoreOptionsExtension.cs" />
     <Compile Include="Internal\DatabaseProviderSelector.cs" />
     <Compile Include="Internal\EnumerableExtensions.cs" />
     <Compile Include="Internal\Graph.cs" />
@@ -220,7 +220,6 @@
     <Compile Include="Metadata\Internal\InternalMetadataBuilder`.cs" />
     <Compile Include="Metadata\Internal\MetadataHelper.cs" />
     <Compile Include="Metadata\ValueGenerated.cs" />
-    <Compile Include="Infrastructure\CoreOptionsExtension.cs" />
     <Compile Include="Infrastructure\DbContextOptions.cs" />
     <Compile Include="Infrastructure\DbContextOptions`.cs" />
     <Compile Include="Infrastructure\IDbContextOptions.cs" />
@@ -248,6 +247,7 @@
     <Compile Include="Metadata\Internal\IFieldMatcher.cs" />
     <Compile Include="Metadata\Internal\IMemberMapper.cs" />
     <Compile Include="ModelBuilder.cs" />
+    <Compile Include="QueryTrackingBehavior.cs" />
     <Compile Include="Query\Annotations\QueryAnnotation.cs" />
     <Compile Include="Query\Annotations\IncludeQueryAnnotation.cs" />
     <Compile Include="EF.cs" />
