+++ src/EntityFramework.Core/Internal/CoreOptionsExtension.cs
@@ -2,10 +2,11 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using JetBrains.Annotations;
+using Microsoft.Data.Entity.Infrastructure;
 using Microsoft.Data.Entity.Metadata;
 using System;
 
-namespace Microsoft.Data.Entity.Infrastructure
+namespace Microsoft.Data.Entity.Internal
 {
     /// <summary>
     ///     Used to store the options specified via <see cref="DbContextOptionsBuilder"/> that are applicable to all databases.
