+++ src/EntityFramework.Core/QueryTrackingBehavior.cs
@@ -1,7 +1,9 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-namespace Microsoft.Data.Entity.ChangeTracking
+using Microsoft.Data.Entity.ChangeTracking;
+
+namespace Microsoft.Data.Entity
 {
     /// <summary>
     ///     Indicates how the results of a query are tracked by the <see cref="ChangeTracker"/>.
