+++ test/EntityFramework.Core.Tests/DbContextOptionsTest.cs
@@ -3,6 +3,7 @@
 
 using System.Linq;
 using Microsoft.Data.Entity.Infrastructure;
+using Microsoft.Data.Entity.Internal;
 using Microsoft.Data.Entity.Metadata;
 using Xunit;
